refactor(star-detection): extract the Star signal copy to StarSignalCopy
Pure move plus re-point, no behavior change: the same 3160 tests pass, and the
moved bodies diff clean against their originals apart from four cross-class
cref qualifications in doc comments.

The block's whole difficulty is that its three surfaces must not repeat or
contradict one another -- the row carries the numbers, the tooltip the
derivation, and the paragraph is capped at a diagnosis plus at most one
instruction. That invariant spans all three, so it can only be reviewed if they
are read together, and they were not: the paragraph and tooltip sat in a
4,000-line ViewModel and the row on the summary model. The tell was
OptimizationSummaryTests reaching into the VM for a static.

Moved: DescribeExposureRecommendation, RemedyFor, DescribeExposureDerivation,
DescribeSweepCost, FormatExposureSeconds, and OptimizationSummary.ExposureText
(now DescribeExposureRow). The MyMessageBox confirmation bodies deliberately
stay on the VM beside DescribeReoptimizeAtBinning: they are typeset to a hard
line budget for a dialog that does not wrap, which is a different constraint,
and they belong with their sibling prompt.

The tests move with them, whole and unedited, into StarSignalCopyTests -- the
block-visibility and BuildCurrentSummary carry-over cases included, since they
share the same fixtures and the same concern. Only the three ExposureText_*
method names are renamed to ExposureRow_*, to match the member they now call.
